requirements.txt
fastmcp>=2.10.6
httpx>=0.28.1
ipykernel>=6.30.0
langchain>=0.3.27
langchain-core>=0.3.72
langchain-mcp-adapters>=0.1.9
langchain-openai>=0.3.28
langgraph-supervisor>=0.0.28
langgraph>=0.5.4
numpy>=2.3.0
python-dotenv>=1.1.1
sqlalchemy>=2.0.41
